+++ CRM/CivirulesConditions/Form/Contribution/xthContributionLast.php
@@ -1,19 +1,19 @@
 <?php
 
 /**
- * Class for CiviRules Condition xth Contribution Form
+ * Class for CiviRules Condition xth Contribution in the last interval Form
  *
- * @author Erik Hommel (CiviCooP) <erik.hommel@civicoop.org>
- * @date 12 Nov 2018
- * @funded by Amnesty International Vlaanderen
+ * @author Sandor Semsey <sandor@es-progress.hu>
+ * @date 16 Feb 2023
  * @license AGPL-3.0
  */
-class CRM_CivirulesConditions_Form_Contribution_xthContribution extends CRM_CivirulesConditions_Form_Form {
+class CRM_CivirulesConditions_Form_Contribution_xthContributionLast extends CRM_CivirulesConditions_Form_Form {
 
   /**
    * Overridden parent method to build form
    *
    * @access public
+   * @throws \Exception
    */
   public function buildQuickForm() {
     $this->add('hidden', 'rule_condition_id');
@@ -23,10 +23,28 @@ class CRM_CivirulesConditions_Form_Contribution_xthContribution extends CRM_Civi
     $this->add('text', 'number_contributions', ts('Number of Contributions'), [], TRUE);
     $this->addRule('number_contributions','Number of Contributions must be a whole number','numeric');
     $this->addRule('number_contributions','Number of Contributions must be a whole number','nopunctuation');
+    $status = CRM_Civirules_Utils_OptionGroup::getActiveValues(CRM_Civirules_Utils::getOptionGroupIdWithName('contribution_status'));
+    $this->add('select', 'contribution_status', ts('Contribution status'), $status, TRUE, ['multiple' => 'multiple', 'class' => 'crm-select2']);
+    $this->add('text', 'interval', ts('in the last'), [], TRUE);
+    $this->add('select', 'interval_unit', ts('interval'), self::getIntervalUnits(), TRUE);
     $this->addButtons([
       ['type' => 'next', 'name' => ts('Save'), 'isDefault' => TRUE],
       ['type' => 'cancel', 'name' => ts('Cancel')]
     ]);
+  }
+
+  /**
+   * Return available time interval units
+   *
+   * @return array
+   */
+  public static function getIntervalUnits(): array
+  {
+    return [
+      'days' => ts('days'),
+      'months' => ts('months'),
+      'years' => ts('years'),
+    ];
   }
 
   /**
@@ -47,6 +65,15 @@ class CRM_CivirulesConditions_Form_Contribution_xthContribution extends CRM_Civi
     if (!empty($data['operator'])) {
       $defaultValues['operator'] = $data['operator'];
     }
+    if (!empty($data['contribution_status'])) {
+      $defaultValues['contribution_status'] = $data['contribution_status'];
+    }
+    if (!empty($data['interval'])) {
+      $defaultValues['interval'] = $data['interval'];
+    }
+    if (!empty($data['interval_unit'])) {
+      $defaultValues['interval_unit'] = $data['interval_unit'];
+    }
     return $defaultValues;
   }
 
@@ -56,7 +83,7 @@ class CRM_CivirulesConditions_Form_Contribution_xthContribution extends CRM_Civi
    * @access public
    */
   public function addRules() {
-    $this->addFormRule(array('CRM_CivirulesConditions_Form_Contribution_xthContribution', 'validateCompareZero'));
+    $this->addFormRule(array('CRM_CivirulesConditions_Form_Contribution_xthContributionLast', 'validateCompareZero'));
   }
 
   /**
@@ -84,6 +111,9 @@ class CRM_CivirulesConditions_Form_Contribution_xthContribution extends CRM_Civi
    * @access public
    */
   public function postProcess() {
+    $data['contribution_status'] = $this->_submitValues['contribution_status'];
+    $data['interval'] = $this->_submitValues['interval'];
+    $data['interval_unit'] = $this->_submitValues['interval_unit'];
     $data['number_contributions'] = $this->_submitValues['number_contributions'];
     $data['operator'] = $this->_submitValues['operator'];
     $data['financial_type'] = $this->_submitValues['financial_type'];