+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## Version 2.46 (not yet released)
+## Version 2.46
 
 * Fix action AddToCase -> error on api3 getvalue with limit => 0, changed to 1
 * Add Activity action: remember send_email "no" selection with !191
@@ -8,6 +8,10 @@
 * Added hours and minutes option to the Activity Scheduled Date trigger (#191)
 * When trigger is scheduled before the activity date/time it triggers now _after_ the trigger time is reached (See #191)
 * Add `Contribution Source` and `Contribution Is Pay Later` conditions  (See !194)
+* Add `Contact is (not) of Type(s)` condition (See !195)
+* Add condition `xth Contribution in Last Time Interval` (see !196)
+* Fix #163: retrieve custom field if needed for activity add action when delayed date in custom field is used
+
 
 ## Version 2.45
 * add xmlMenu function to civirules.php (mixin issue with older CiviCRM versions)